+++ core/src/main/scala/com/nvidia/spark/rapids/tool/profiling/Profiler.scala
@@ -20,7 +20,6 @@ import java.util.concurrent.{ConcurrentLinkedQueue, Executors, ThreadPoolExecuto
 
 import scala.collection.JavaConverters._
 import scala.collection.mutable.{ArrayBuffer, HashMap}
-import scala.io.Source
 import scala.util.control.NonFatal
 
 import com.nvidia.spark.rapids.ThreadFactoryBuilder
@@ -130,36 +129,13 @@ class Profiler(hadoopConf: Configuration, appArgs: ProfileArgs, enablePB: Boolea
       Profiler.DRIVER_LOG_NAME, numOutputRows, true)
 
     try {
-      val unsupportedDrivers = processDriverLog(driverLogInfos)
+      val driverLogProcessor = new DriverLogProcessor(driverLogInfos)
+      val unsupportedDrivers = driverLogProcessor.processDriverLog()
       profileOutputWriter.write(s"Unsupported operators in driver log",
         unsupportedDrivers)
     } finally {
       profileOutputWriter.close()
     }
-  }
-
-  def processDriverLog(driverlogPath: String): Seq[DriverLogUnsupportedOperators] = {
-    val source = Source.fromFile(driverlogPath)
-    // Create a map to store the counts for each operator and reason
-    var countsMap = Map[(String, String), Int]().withDefaultValue(0)
-    try {
-      // Process each line in the file
-      for (line <- source.getLines()) {
-        if (line.contains("cannot run on GPU") &&
-          !line.contains("not all expressions can be replaced")) {
-          val operatorName = line.split("<")(1).split(">")(0)
-          val reason = line.split("because")(1).trim()
-          val key = (operatorName, reason)
-          countsMap += key -> (countsMap(key) + 1)
-        }
-      }
-    } catch {
-      case e: Exception =>
-        logError(s"Unexpected exception processing driver log: $driverlogPath", e)
-    } finally {
-      source.close()
-    }
-    countsMap.map(x => DriverLogUnsupportedOperators(x._1._1, x._2, x._1._2)).toSeq
   }
 
   private def errorHandler(error: Throwable, path: EventLogInfo) = {
